cfn/tenant-cfn/requirements.txt
pyyaml 
python-docx
